% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgoslin.R
\name{isValidLipidName}
\alias{isValidLipidName}
\title{Check lipid name.}
\usage{
isValidLipidName(lipidName)
}
\arguments{
\item{lipidName}{The lipid name to check.}
}
\value{
TRUE if the lipidName could be parsed, FALSE otherwise.
}
\description{
\code{isValidLipidName} checks the provided lipid name against the built-in
grammars.
Will return FALSE if none of the parsers was able to parse the provided name
successfully. Will stop execution via \code{stop} if non character input is 
detected.
}
\examples{
isValidLipidName("PC 32:1")
isValidLipidName("PC(32:1)")
isValidLipidName("PCX(32:1)")
}
