% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgoslin.R
\name{listAvailableGrammars}
\alias{listAvailableGrammars}
\title{Return the list of grammars supported by goslin.}
\usage{
listAvailableGrammars()
}
\value{
the list of grammars
}
\description{
\code{listAvailableGrammars} returns the list of grammars that the
underlying cppgoslin library supports.
}
\examples{
listAvailableGrammars()
}
