% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgoslin.R
\name{parseLipidNamesWithGrammar}
\alias{parseLipidNamesWithGrammar}
\title{Parse multiple lipid names with the provided grammar and return a data frame
with the results.}
\usage{
parseLipidNamesWithGrammar(lipidNames, grammar)
}
\arguments{
\item{lipidNames}{The vector of lipid names to parse.}

\item{grammar}{The grammar to use. One of "Goslin", "GoslinFragments",
"SwissLipids", "LipidMaps", "HMDB".}
}
\value{
Data frame where each row reports the parsing result of each element
in lipidNames.
}
\description{
\code{parseLipidNamesWithGrammar} reads the provided lipid names vector and
parses it against the given grammar. It returns structural information as a
data frame. Will return a cell with the "Grammar" column set to
"NOT_PARSEABLE" if none of the parsers was able to parse the provided name
successfully. If any error was raised, returns an empty data frame.
}
\examples{
parseLipidNamesWithGrammar(c("PC 32:1","LPC 34:1","TG(18:1_18:0_16:1)"),
"Goslin")
}
